requirements.txt
git+https://github.com/mlflow/mlflow.git@databricks-trace-archival
https://ml-team-public-read.s3.us-west-2.amazonaws.com/wheels/mlflow/dev/edwin.he/databricks_zerobus-0.0.16-py3-none-any.whl
fastapi
uvicorn
pandas
pydantic
protobuf
opentelemetry-proto
